Reto 1/microservicios_django/api_usuarios/requirements.txt
# Para api_usuarios
Django>=4.2.0
django-ninja>=1.0.1
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
pandas>=2.1.0
requests>=2.31.0